+++ src/api/files.py
@@ -1,39 +1,60 @@
 """
-File listing and search API endpoints.
+Internal file listing and search handlers (cookie auth).
 
-Provides server-side file-level views over the chunk-based OpenSearch index.
-Uses terms aggregation with in-memory sort/slice (v1).
+Serves the unversioned internal/UI routes GET /files and GET /files/search
+using FileServiceV2 (composite-aggregation cursor pagination). Pass `after_key`
+(JSON-encoded) to advance pages.
+
 """
 
-from fastapi import Depends, Query
+import json
+
+from fastapi import Depends, HTTPException, Query
 from fastapi.responses import JSONResponse
 
-from dependencies import get_current_user, get_session_manager
+from dependencies import get_current_user, get_file_service_v2
 from session_manager import User
 from utils.logging_config import get_logger
 
 logger = get_logger(__name__)
 
 
-def _get_file_service(session_manager=Depends(get_session_manager)):
-    from services.file_service import FileService
+def _parse_after_key(after_key: str | None) -> dict | None:
+    """Parse a JSON-encoded composite cursor.
 
-    return FileService(session_manager=session_manager)
+    Returns None when after_key is absent.
+    Raises HTTPException 400 when the value is present but is not valid JSON
+    or does not decode to a dict (e.g. a bare string or number is invalid).
+    """
+    if not after_key:
+        return None
+    try:
+        parsed = json.loads(after_key)
+    except (json.JSONDecodeError, ValueError) as err:
+        raise HTTPException(status_code=400, detail="after_key is not valid JSON") from err
+    if not isinstance(parsed, dict):
+        raise HTTPException(status_code=400, detail="after_key must be a JSON object")
+    return parsed
 
 
 async def list_files(
-    page: int = Query(1, ge=1, description="Page number"),
-    page_size: int = Query(25, ge=1, le=100, description="Items per page"),
+    page: int = Query(
+        1, ge=1, description="Page number (for display only; navigation uses after_key cursor)"
+    ),
+    page_size: int = Query(25, ge=1, le=500, description="Items per page"),
     sort_by: str = Query("filename", description="Sort field"),
     sort_order: str = Query("asc", pattern="^(asc|desc)$", description="Sort order"),
     connector_type: str | None = Query(None, description="Filter by connector type"),
     mimetype: str | None = Query(None, description="Filter by MIME type"),
     owner: str | None = Query(None, description="Filter by owner"),
     search: str | None = Query(None, description="Search filename"),
-    file_service=Depends(_get_file_service),
+    after_key: str | None = Query(None, description="Composite pagination cursor (JSON-encoded)"),
+    file_service=Depends(get_file_service_v2),
     user: User = Depends(get_current_user),
 ):
-    """List ingested files with pagination, filtering, and sorting."""
+    """List ingested files with composite-aggregation pagination, filtering, and sorting."""
+    parsed_after_key = _parse_after_key(after_key)
+
     try:
         result = await file_service.list_files(
             user_id=user.user_id,
@@ -46,16 +67,17 @@ async def list_files(
             mimetype=mimetype,
             owner=owner,
             search=search,
+            after_key=parsed_after_key,
         )
         return JSONResponse(result)
     except Exception as e:
-        logger.error("Failed to list files", error=str(e))
+        logger.error("Failed to list files (v2)", error=str(e))
         from utils.opensearch_utils import AUTH_ERROR_MESSAGE, is_opensearch_auth_error
 
         if is_opensearch_auth_error(e):
             return JSONResponse({"error": AUTH_ERROR_MESSAGE}, status_code=401)
         return JSONResponse(
-            {"error": "Failed to list files", "detail": str(e)},
+            {"error": "Failed to list files"},
             status_code=500,
         )
 
@@ -63,14 +85,17 @@ async def list_files(
 async def search_files(
     q: str = Query(..., min_length=1, description="Search query"),
     page: int = Query(1, ge=1, description="Page number"),
-    page_size: int = Query(25, ge=1, le=100, description="Items per page"),
+    page_size: int = Query(25, ge=1, le=500, description="Items per page"),
     connector_type: str | None = Query(None, description="Filter by connector type"),
     mimetype: str | None = Query(None, description="Filter by MIME type"),
     owner: str | None = Query(None, description="Filter by owner"),
-    file_service=Depends(_get_file_service),
+    after_key: str | None = Query(None, description="Composite pagination cursor (JSON-encoded)"),
+    file_service=Depends(get_file_service_v2),
     user: User = Depends(get_current_user),
 ):
-    """Search files by name with fuzzy/partial matching."""
+    """Search files by name with fuzzy/partial matching (composite pagination)."""
+    parsed_after_key = _parse_after_key(after_key)
+
     try:
         result = await file_service.search_files(
             user_id=user.user_id,
@@ -81,15 +106,16 @@ async def search_files(
             connector_type=connector_type,
             mimetype=mimetype,
             owner=owner,
+            after_key=parsed_after_key,
         )
         return JSONResponse(result)
     except Exception as e:
-        logger.error("Failed to search files", error=str(e))
+        logger.error("Failed to search files (v2)", error=str(e))
         from utils.opensearch_utils import AUTH_ERROR_MESSAGE, is_opensearch_auth_error
 
         if is_opensearch_auth_error(e):
             return JSONResponse({"error": AUTH_ERROR_MESSAGE}, status_code=401)
         return JSONResponse(
-            {"error": "Failed to search files", "detail": str(e)},
+            {"error": "Failed to search files"},
             status_code=500,
         )
