+++ src/api/v1/files.py
@@ -2,9 +2,8 @@
 Public API v1 Files endpoints.
 
 Provides offset-paginated file listing over the ingested knowledge base
-(GET /v1/files/getAll). Uses API-key authentication and calls the shared
-FileServiceV2 directly. Cursor-paginated variants live in api/v2/files.py
-(list_files_public / search_files_public, served at /v2/files).
+(GET /v1/files/getAll). Uses API-key authentication and calls FileService
+(v1, offset pagination) directly. .
 """
 
 from fastapi import Depends, Query
